src/twitter/requirements.txt
aiohttp==3.7.4
antlr4-python3-runtime==4.9.1
appdirs==1.4.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
identify==1.5.13
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.1.0
pluggy==0.13.1
praw==7.1.4
prawcore==1.5.0
pre-commit==2.10.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyglet==1.5.15
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.0.0
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
sympy==1.7.1
toml==0.10.2
tweepy==3.10.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.3
virtualenv==20.4.2
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.6.3
